Changed labels to meet requirements
- mg/L -> in milligrams per liter

- kg*6^10 -> in thousands of tons

- Changed Y-axis labels to not be canvas since canvas doesn't allow for
a linebreak using \n or <br /> and the label changes make the entire
label too large to be on one line. Flot doesn't create a line break when
needed. Seemingly, the only way to get around this is to make the plots
taller to fit the text. I tried getting the text down to a font size 8
but then it's way too small to read 

**This could break functionality down the line if we need to export
canvas!!