Added a stream flow time series select which is disabled unless Stream Flow is checked.
Constituent and water quality data type are now disabled if water
quality is unchecked. Fields are cleared as well

Using a form tag in the markup so we can more easily collect the
parameters.

Clear Filters now resets the whole form including dates.

Set minDate and maxDate on endDate and startDate respectively when the
startDate or endDate is updated. Clearing the form reset the minDate and
maxDate back to the defaults.