Defaulted the Mississipi page selects as requested.
Added code to toggle the sites layer on and off via the Toggle Sites
button. Changed the way the flow lines are added to the map. These are
now added at page rendering and is done before the sites layers are
added. Using setVisibility on layers to turn them on/off.