Added the Alaska inset map. Added the code to the map.js file since both maps share layers and other settings.
The map height was increased and the center and resolution adjusted so
that the US 48 takes up more of the space while still leaving room for
the Alaska inset map. 

At the moment, I am using the Street Map for the Alaska inset. Waiting
for streams and Alaska outline shapefiles.